teach the parser how to handle schema value inside the opts key
