+++ test/analyzer/analyzer.from.test.js
@@ -30,44 +30,5 @@ describe('Analyzer ::', function() {
         return done();
       });
     });
-
-    it('should generate a valid group for select when FROM is used with a SCHEMA', function(done) {
-      var tokens = tokenize({
-        select: ['title', 'author', 'year'],
-        from: { table: 'books', schema: 'foo' }
-      });
-
-      Analyzer({
-        tokens: tokens
-      })
-      .exec(function(err, result) {
-        assert(!err);
-
-        assert.deepEqual(result,  [
-          [
-            { type: 'IDENTIFIER', value: 'SELECT' },
-            { type: 'VALUE', value: 'title' }
-          ],
-          [
-            { type: 'IDENTIFIER', value: 'SELECT' },
-            { type: 'VALUE', value: 'author' }
-          ],
-          [
-            { type: 'IDENTIFIER', value: 'SELECT' },
-            { type: 'VALUE', value: 'year' }
-          ],
-          [
-            { type: 'IDENTIFIER', value: 'SCHEMA' },
-            { type: 'VALUE', value: 'foo' }
-          ],
-          [
-            { type: 'IDENTIFIER', value: 'FROM' },
-            { type: 'VALUE', value: 'books' }
-          ]
-        ]);
-
-        return done();
-      });
-    });
   });
 });
