+++ test/analyzer/analyzer.opts.schema.test.js
@@ -1,0 +1,49 @@
+var Analyzer = require('../../index').analyzer;
+var tokenize = require('../support/tokenize');
+var assert = require('assert');
+
+describe('Analyzer ::', function() {
+  describe('OPTS', function() {
+    it('should generate a valid group when a SCHEMA is used', function(done) {
+      var tokens = tokenize({
+        select: ['title', 'author', 'year'],
+        from: 'books',
+        opts: {
+          schema: 'foo'
+        }
+      });
+
+      Analyzer({
+        tokens: tokens
+      })
+      .exec(function(err, result) {
+        assert(!err);
+
+        assert.deepEqual(result,  [
+          [
+            { type: 'IDENTIFIER', value: 'SELECT' },
+            { type: 'VALUE', value: 'title' }
+          ],
+          [
+            { type: 'IDENTIFIER', value: 'SELECT' },
+            { type: 'VALUE', value: 'author' }
+          ],
+          [
+            { type: 'IDENTIFIER', value: 'SELECT' },
+            { type: 'VALUE', value: 'year' }
+          ],
+          [
+            { type: 'IDENTIFIER', value: 'FROM' },
+            { type: 'VALUE', value: 'books' }
+          ],
+          [
+            { type: 'IDENTIFIER', value: 'SCHEMA' },
+            { type: 'VALUE', value: 'foo' }
+          ]
+        ]);
+
+        return done();
+      });
+    });
+  });
+});
